fix: scatter dtype mismatch under torch.compile + bfloat16
F.normalize promotes bfloat16 → float32 internally (norm uses float32 for
stability), so top_k_weights_flat ends up float32 while expert_weights =
zeros_like(raw_scores) stays bfloat16 → scatter() dtype mismatch crash
under torch.compile with Qwen2 bf16 training.

Fix: .to(expert_weights.dtype) / .to(probs.dtype) on the scatter src.
Affects deepseek.py and standard.py (same pattern).

+++ src/routers/deepseek.py
@@ -93,7 +93,7 @@ class DeepSeekRouter(BaseRouter):
 
         expert_weights = torch.zeros_like(raw_scores_flat)
         expert_weights = expert_weights.scatter(
-            1, top_k_indices_flat, top_k_weights_flat
+            1, top_k_indices_flat, top_k_weights_flat.to(expert_weights.dtype)
         )
 
         if self.training and record_usage:
